sdk/debug/lost-context-simulator-test.html

<!--
Copyright (c) 2009 The Chromium Authors. All rights reserved.
Use of this source code is governed by a BSD-style license that can be
found in the LICENSE file.
 -->
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
  <head>
    <title>WebGL Lost Context Wrapper Test</title>
    <link rel="stylesheet" href="../tests/resources/js-test-style.css"/>
    <script src="../tests/resources/js-test-pre.js"></script>
    <script src="../tests/conformance/resources/webgl-test.js"> </script>
    <script src="webgl-debug.js"> </script>
</head>
<body>
<canvas id="example" width="2" height="2" style="width: 40px; height: 40px;"></canvas>
<div id="description"></div>
<div id="console"></div>
<script>
debug("Tests the context lost simulator.");
debug("");
var gl = create3DContext(document.getElementById("example"));
//gl = WebGLDebugUtils.makeDebugContext(gl);
//WebGLDebugUtils.resetToInitialState(gl);
gl = WebGLDebugUtils.makeLostContextSimulatingContext(gl);
gl.registerOnContextRestoredListener(afterContextRestored);

assertMsg(gl, "Got 3d context.");

var vertexObject = gl.createBuffer();
gl.bindBuffer(gl.ARRAY_BUFFER, vertexObject);
glErrorShouldBe(gl, gl.NO_ERROR, "no errors from setup");
gl.loseContext();
glErrorShouldBe(gl, gl.CONTEXT_LOST_WEBGL, "error should be lost context");
shouldBeNull("gl.createBuffer()");
gl.restoreContext();
function afterContextRestored() {
  glErrorShouldBe(gl, gl.NO_ERROR, "no errors after context restored");
  gl.bindBuffer(gl.ARRAY_BUFFER, vertexObject);
  glErrorShouldBe(gl, gl.INVALID_OPERATION, "invalid operation using a resource from lost context");
  vertexObject = gl.createBuffer();
  gl.bindBuffer(gl.ARRAY_BUFFER, vertexObject);
  glErrorShouldBe(gl, gl.NO_ERROR, "okay to use new context");

  successfullyParsed = true;
  var epilogue = document.createElement("script");
  epilogue.onload = finish;
  epilogue.src = "../tests/resources/js-test-post.js";
  document.body.appendChild(epilogue);
}

function finish() {
    if (window.nonKhronosFrameworkNotifyDone) {
        window.nonKhronosFrameworkNotifyDone();
    }
}

</script>
</body>
</html>
